backend/canvas_service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
PyJWT[crypto]>=2.8.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
supabase>=2.4.0
python-multipart>=0.0.9
websockets>=12.0
